+++ app/src/main/java/org/pathfinderfr/app/ItemDetailActivity.java
@@ -54,6 +54,7 @@ public class ItemDetailActivity extends AppCompatActivity {
             // using a fragment transaction.
             Bundle arguments = new Bundle();
             arguments.putLong(ItemDetailFragment.ARG_ITEM_ID, getIntent().getLongExtra(ItemDetailFragment.ARG_ITEM_ID, 0));
+            arguments.putLong(ItemDetailFragment.ARG_ITEM_SEL_CHARACTER, getIntent().getLongExtra(ItemDetailFragment.ARG_ITEM_SEL_CHARACTER, 0));
             arguments.putString(ItemDetailFragment.ARG_ITEM_FACTORY_ID, getIntent().getStringExtra(ItemDetailFragment.ARG_ITEM_FACTORY_ID));
             arguments.putBoolean(ItemDetailFragment.ARG_ITEM_SHOWDETAILS, getIntent().getBooleanExtra(ItemDetailFragment.ARG_ITEM_SHOWDETAILS, showDetails));
             arguments.putString(ItemDetailFragment.ARG_ITEM_MESSAGE, getIntent().getStringExtra(ItemDetailFragment.ARG_ITEM_MESSAGE));